# swagger-idpay.yaml
openapi: 3.0.3
info:
  title: IDPAY ITN IO API
  version: '1.0'
  contact:
    name: PagoPA S.p.A.
    email: cstar@pagopa.it
  description: |-
    IDPAY ITN IBAN IO

    IDPAY ITN Onboarding Workflow IO

    IDPAY ITN Timeline IO

    IDPAY ITN Wallet IO

    IDPAY ITN Payment QRCODE IO

    IDPAY ITN Payment IO
servers:
  - description: Development Test
    url: 'https://api-io.dev.cstar.pagopa.it/idpay-itn/wallet'
    x-internal: true
  - description: User Acceptance Test
    url: 'https://api-io.uat.cstar.pagopa.it/idpay-itn/wallet'
    x-internal: true
security:
  - bearerAuth: []
...
paths:
...
  '/onboarding/service/{serviceId}':
    get:
      tags:
        - onboarding
      summary: >-
        ENG: Retrieves the initiative ID starting from the corresponding service
        ID - IT: Ritrova l'identificativo dell'iniziativa a partire
        dall'idetificativo del service
      description: Retrieves the initiative ID starting from the corresponding service ID
      operationId: getInitiativeData
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - name: serviceId
          in: path
          description: 'ENG: The service ID - IT: Identificativo del service'
          required: true
          schema:
            type: string
            maxLength: 50
            pattern: '$ ^[a-zA-Z0-9]+$'
        - name: Accept-Language
          in: header
          description: 'ENG: Language - IT: Lingua'
          schema:
            type: string
            pattern: '^[ -~]{2,5}$'
            minLength: 2
            maxLength: 5
            example: it-IT
            default: it-IT
          required: true
      responses:
        '200':
          description: Get successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiativeDataDTO'
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiativeErrorDTO'
              example:
                code: INITIATIVE_INVALID_REQUEST
                message: Something went wrong handling the request
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '401':
          description: Authentication failed
          content:
            application/json: {}
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '404':
          description: The requested initiative was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiativeErrorDTO'
              example:
                code: INITIATIVE_NOT_FOUND
                message: Initiative not found
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '429':
          description: Too many Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiativeErrorDTO'
              example:
                code: INITIATIVE_TOO_MANY_REQUESTS
                message: Too many requests
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '500':
          description: Server ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiativeErrorDTO'
              example:
                code: INITIATIVE_GENERIC_ERROR
                message: Application error
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
...
  /onboarding/:
    put:
      tags:
        - onboarding
      summary: >-
        ENG: Saving mail and self-declarations - IT: Salvataggio della mail e
        delle autodichiarazioni
      description: Saving mail and self-declarations
      operationId: saveOnboarding
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - name: Accept-Language
          in: header
          description: 'ENG: Language - IT: Lingua'
          schema:
            type: string
            pattern: '^[ -~]{2,5}$'
            minLength: 2
            maxLength: 5
            example: it-IT
            default: it-IT
          required: true
      requestBody:
        description: >-
          ENG: Unique identifier of the subscribed initiative, flag for PDND
          acceptation and the list of accepted self-declared criteria - IT:
          Identificativo univoco dell'iniziativa sottoscritta, elenco dei
          criteri autodichiarati accettati,
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingDTO'
      responses:
        '202':
          description: Accepted - Request Taken Over
          content:
            application/json: {}
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingErrorDTO'
              example:
                code: ONBOARDING_INVALID_REQUEST
                message: Something went wrong handling the request
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '401':
          description: Authentication failed
          content:
            application/json: {}
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '403':
          description: This onboarding is forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingErrorDTO'
              example:
                code: ONBOARDING_BUDGET_EXHAUSTED
                message: Budget exhausted for initiative XXXXX
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '404':
          description: The requested resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingErrorDTO'
              example:
                code: ONBOARDING_INITIATIVE_NOT_FOUND
                message: Cannot find initiative XXXXX
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '429':
          description: Too many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingErrorDTO'
              example:
                code: ONBOARDING_TOO_MANY_REQUESTS
                message: Too many requests
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '500':
          description: Server ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingErrorDTO'
              example:
                code: ONBOARDING_GENERIC_ERROR
                message: An error occurred in the microservice admissibility
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
  '/onboarding/{initiativeId}/status':
    get:
      tags:
        - onboarding
      summary: >-
        ENG: Returns the actual onboarding status along with the date on which
        that status changed and, if present, the date upon which the onboarding
        successfully went through - IT: Ritorna lo stato attuale dell'adesione
        insieme alla data in cui quello stato è cambiato e, se presente, la data
        in cui l'adesione è avvenuta con successo
      description: >-
        Returns the actual onboarding status along with the date on which that
        status changed and, if present, the date upon which the onboarding
        successfully went through
      operationId: onboardingStatus
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - name: Accept-Language
          in: header
          description: 'ENG: Language - IT: Lingua'
          schema:
            type: string
            pattern: '^[ -~]{2,5}$'
            minLength: 2
            maxLength: 5
            example: it-IT
            default: it-IT
          required: true
        - name: initiativeId
          in: path
          description: 'ENG: The initiative ID - IT: Identificativo dell''iniziativa'
          required: true
          schema:
            type: string
            maxLength: 24
            pattern: '$ ^[a-zA-Z0-9]+$'
      responses:
        '200':
          description: Check successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingStatusDTO'
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingErrorDTO'
              example:
                code: ONBOARDING_INVALID_REQUEST
                message: Something went wrong handling the request
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '401':
          description: Authentication failed
          content:
            application/json: {}
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '404':
          description: The requested resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingErrorDTO'
              example:
                code: ONBOARDING_USER_NOT_ONBOARDED
                message: The current user is not onboarded on initiative XXXXX
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '429':
          description: Too many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingErrorDTO'
              example:
                code: ONBOARDING_TOO_MANY_REQUESTS
                message: Too many requests
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '500':
          description: Server ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingErrorDTO'
              example:
                code: ONBOARDING_GENERIC_ERROR
                message: An error occurred in the microservice admissibility
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
...
components:
  schemas:
...
    SelfConsentDTO:
      oneOf:
        - $ref: '#/components/schemas/SelfConsentBoolDTO'
        - $ref: '#/components/schemas/SelfConsentMultiDTO'
        - $ref: '#/components/schemas/SelfConsentTextDTO'
        - $ref: '#/components/schemas/SelfConsentMultiTypeDTO'
...
    OnboardingStatusDTO:
      title: OnboardingStatusDTO
      type: object
      required:
        - status
        - statusDate
      properties:
        status:
          enum:
            - ACCEPTED_TC
            - ON_EVALUATION
            - ONBOARDING_KO
            - ELIGIBLE_KO
            - ONBOARDING_OK
            - UNSUBSCRIBED
            - INVITED
            - DEMANDED
            - SUSPENDED
          type: string
          description: >-
            ENG: Actual status of the citizen onboarding for an initiative - IT:
            Stato attuale del cittadino rispetto ad un'iniziativa
        statusDate:
          type: string
          format: date-time
          description: >-
            ENG: Date on which the status changed to the current one - IT: Data
            in cui lo stato è cambiato allo stato attuale
        onboardingOkDate:
          type: string
          format: date-time
          description: >-
            ENG: Date on which the onboarding successfully went through - IT:
            Data in cui l'adesione è avvenuta con successo
...
    SelfConsentBoolDTO:
      type: object
      required:
        - _type
        - code
        - accepted
      properties:
        _type:
          type: string
          enum:
            - boolean
          description: 'ENG: Self-consent boolean type - IT: Auto consenso di tipo boolean'
        code:
          type: string
          description: 'ENG: Self-consent code - IT: Codice dell''auto consenso'
        accepted:
          type: boolean
          description: >-
            ENG: Indicates whether the self-consent is accepted or not - IT:
            Indica se l'auto consenso è accettato o no
    SelfConsentMultiDTO:
      type: object
      required:
        - _type
        - code
        - value
      properties:
        _type:
          type: string
          enum:
            - multi
          description: 'ENG: Self-consent value type - IT: Auto consenso di tipo multipli'
        code:
          type: string
          description: 'ENG: Self-consent code - IT: Codice dell''auto consenso'
        value:
          type: string
          example: '0, 1'
          description: >-
            ENG: Indicates self-consent values - IT: Indica i valori per gli
            auto consensi
    SelfConsentTextDTO:
      type: object
      required:
        - _type
        - code
        - value
      properties:
        _type:
          type: string
          enum:
            - text
          description: >-
            ENG: Self-consent value type free text - IT: Auto consenso di tipo
            testo libero
        code:
          type: string
          description: 'ENG: Self-consent code - IT: Codice dell''auto consenso'
        value:
          type: string
          description: >-
            ENG: Indicates self-consent values - IT: Indica i valori per gli
            auto consensi
    SelfConsentMultiTypeDTO:
      type: object
      required:
        - _type
        - code
        - value
      properties:
        _type:
          type: string
          enum:
            - multi_consent
          description: >-
            ENG: ISEE Self-consent value type - IT: Auto consenso di tipo
            multipli ISEE
        code:
          type: string
          enum:
            - isee
          description: 'ENG: Self-consent code - IT: Tipologia di multiconsenso'
        value:
          type: string
          example: '1, 2, 3'
          description: >-
            ENG: Indicates self-consent values - IT: Indica i valori per gli
            auto consensi
    InitiativeDataDTO:
      type: object
      required:
        - initiativeId
        - initiativeName
        - description
        - organizationId
        - organizationName
        - tcLink
        - privacyLink
      properties:
        initiativeId:
          type: string
          description: 'ENG: The initiative ID - IT: Identificativo dell''iniziativa'
        initiativeName:
          type: string
          description: 'ENG: Initiative''s name - IT: Nome dell''iniziativa'
          example: Bonus Elettrodomestici
        description:
          type: string
          example: Descrizione dell'iniziativa
          description: 'ENG: Initiative''s description - IT: Descrizione dell''iniziativa'
        organizationId:
          type: string
          description: >-
            ENG: Id of the organization that created the initiative - IT:
            Identificativo dell'organizzazione che ha creato l'iniziativa
        organizationName:
          type: string
          description: >-
            ENG: Name of the organization that created the initiative - IT: Nome
            dell'organizzazione che ha creato l'iniziativa
        tcLink:
          type: string
          description: >-
            ENG: URL that redirects to the terms and conditions - IT: URL che
            porta ai termini e condizioni
        privacyLink:
          type: string
          description: >-
            ENG: URL that redirects to the privacy policy - IT: URL che
            reindirizza all'informativa della privacy
        logoURL:
          type: string
          description: >-
            ENG: URL for the initiative's logo image - IT: URL del logo
            dell'iniziativa
        thumbnailUrl:
          type: string
          description: >-
            ENG: ToS Initiative Image URL - IT: URL dell'immagine
            dell'iniziativa per i ToS
    InitiativeErrorDTO:
      type: object
      properties:
        code:
          type: string
          enum:
            - INITIATIVE_INVALID_LOCALE_FORMAT
            - INITIATIVE_INVALID_REQUEST
            - INITIATIVE_NOT_FOUND
            - INITIATIVE_TOO_MANY_REQUESTS
            - INITIATIVE_GENERIC_ERROR
          description: |-
            'ENG: Error code:
             INITIATIVE_INVALID_LOCALE_FORMAT: Initiative not found due to invalid locale format,
             INITIATIVE_INVALID_REQUEST: Something went wrong handling the request,
             INITIATIVE_NOT_FOUND: Initiative not found,
             INITIATIVE_TOO_MANY_REQUESTS: Too many requests,
             INITIATIVE_GENERIC_ERROR: Application error,
             - IT: Codice di errore:
             INITIATIVE_INVALID_LOCALE_FORMAT: Iniziativa non trovata a causa di format locale non valido,
             INITIATIVE_INVALID_REQUEST: Qualcosa è andato storto durante l'invio della richiesta,
             INITIATIVE_NOT_FOUND: Iniziativa non trovata,
             INITIATIVE_TOO_MANY_REQUESTS: Troppe richieste,
             INITIATIVE_GENERIC_ERROR: Errore generico'
        message:
          type: string
          description: 'ENG: Error message - IT: Messaggio di errore'
    OnboardingErrorDTO:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - ONBOARDING_USER_UNSUBSCRIBED
            - ONBOARDING_PAGE_SIZE_NOT_ALLOWED
            - ONBOARDING_PDND_CONSENT_DENIED
            - ONBOARDING_SELF_DECLARATION_NOT_VALID
            - ONBOARDING_SUSPENSION_NOT_ALLOWED_FOR_USER_STATUS
            - ONBOARDING_READMISSION_NOT_ALLOWED_FOR_USER_STATUS
            - ONBOARDING_INVALID_REQUEST
            - ONBOARDING_USER_NOT_IN_WHITELIST
            - ONBOARDING_INITIATIVE_NOT_STARTED
            - ONBOARDING_INITIATIVE_ENDED
            - ONBOARDING_BUDGET_EXHAUSTED
            - ONBOARDING_INITIATIVE_STATUS_NOT_PUBLISHED
            - ONBOARDING_TECHNICAL_ERROR
            - ONBOARDING_UNSATISFIED_REQUIREMENTS
            - ONBOARDING_GENERIC_ERROR
            - ONBOARDING_USER_NOT_ONBOARDED
            - ONBOARDING_INITIATIVE_NOT_FOUND
            - ONBOARDING_TOO_MANY_REQUESTS
            - ONBOARDING_FAMILY_UNIT_ALREADY_JOINED
            - ONBOARDING_WAITING_LIST
          description: >-
            'ENG: Error code: ONBOARDING_USER_UNSUBSCRIBED: The user has
            unsubscribed from initiative, ONBOARDING_PAGE_SIZE_NOT_ALLOWED: Page
            size not allowed, ONBOARDING_PDND_CONSENT_DENIED: The PDND consent
            was denied by the user for the initiative,
            ONBOARDING_SELF_DECLARATION_NOT_VALID: The self-declaration criteria
            are not valid for the initiative or those inserted by the user do
            not match those required by the initiative,
            ONBOARDING_SUSPENSION_NOT_ALLOWED_FOR_USER_STATUS: It is not
            possible to suspend the user on initiative,
            ONBOARDING_READMISSION_NOT_ALLOWED_FOR_USER_STATUS: It is not
            possible to readmit the user on initiative,
            ONBOARDING_INVALID_REQUEST: Something went wrong handling the
            request, ONBOARDING_USER_NOT_IN_WHITELIST: The current user is not
            allowed to participate to the initiative,
            ONBOARDING_INITIATIVE_NOT_STARTED: The initiative has not yet begun,
            ONBOARDING_INITIATIVE_ENDED: The initiative ended,
            ONBOARDING_BUDGET_EXHAUSTED: Budget exhausted for initiative,
            ONBOARDING_INITIATIVE_STATUS_NOT_PUBLISHED: The initiative is not
            active, ONBOARDING_TECHNICAL_ERROR: A technical error occurred
            during the onboarding process, ONBOARDING_UNSATISFIED_REQUIREMENTS:
            The user does not satisfy the requirements set for the initiative,
            ONBOARDING_GENERIC_ERROR: Application error,
            ONBOARDING_USER_NOT_ONBOARDED: The current user is not onboarded on
            initiative, ONBOARDING_INITIATIVE_NOT_FOUND: Cannot find initiative,
            ONBOARDING_TOO_MANY_REQUESTS: Too many requests - IT: Codice di
            errore: ONBOARDING_USER_UNSUBSCRIBED: L'utente si è disiscritto
            dall'iniziativa, ONBOARDING_PAGE_SIZE_NOT_ALLOWED: Dimensione pagina
            non permessa, ONBOARDING_PDND_CONSENT_DENIED: Il consenso PDND per
            l'iniziativa è stato negato dall'utente,
            ONBOARDING_SELF_DECLARATION_NOT_VALID: I criteri di
            autodichiarazione non sono validi per l'iniziativa o quelli inseriti
            dall'utente non corrispondono con quelli richiesti dall'iniziativa,
            ONBOARDING_SUSPENSION_NOT_ALLOWED_FOR_USER_STATUS: Non è possibile
            sospendere l'utente dall'iniziativa,
            ONBOARDING_READMISSION_NOT_ALLOWED_FOR_USER_STATUS: Non è possibile
            riammettere l'utente all'iniziativa, ONBOARDING_INVALID_REQUEST:
            Qualcosa è andato storto durante l'invio della richiesta,
            ONBOARDING_USER_NOT_IN_WHITELIST: L'utente corrente non ha il
            permesso di partecipare all'iniziativa,
            ONBOARDING_INITIATIVE_NOT_STARTED: L'iniziativa non è ancora
            partita, ONBOARDING_INITIATIVE_ENDED: L'iniziativa è terminata,
            ONBOARDING_BUDGET_EXHAUSTED: Il budget dell'iniziativa è esaurito,
            ONBOARDING_INITIATIVE_STATUS_NOT_PUBLISHED: L'iniziativa non è
            attiva, ONBOARDING_TECHNICAL_ERROR: Si è verificato un errore
            tecnico durante il processo di adesione,
            ONBOARDING_UNSATISFIED_REQUIREMENTS: L'utente non soddisfa i
            requisiti che sono stati richiesti per l'iniziativa,
            ONBOARDING_GENERIC_ERROR: Errore applicativo,
            ONBOARDING_USER_NOT_ONBOARDED: Utente non onboardato all'iniziativa,
            ONBOARDING_INITIATIVE_NOT_FOUND: Iniziativa non trovata,
            ONBOARDING_TOO_MANY_REQUESTS: Troppe richieste'
        message:
          type: string
          description: 'ENG: Error message- IT: Messaggio di errore'
    OnboardingDTO:
      title: OnboardingWebDTO
      type: object
      required:
        - initiativeId
        - confirmedTos
        - pdndAccept
        - selfDeclarationList
      properties:
        initiativeId:
          type: string
          description: >-
            ENG: Unique identifier of the subscribed initiative - IT:
            Identificativo univoco dell'iniziativa sottoscritta
          maxLength: 24
          pattern: '$ ^[a-zA-Z0-9]+$'
        confirmedTos:
          type: boolean
          description: >-
            'ENG: Indicates whether the user has correctly accepted the Terms
            and Conditions of the initiative - IT: Indica se l'utente ha
            correttamente accettato i Termini e Condizioni dell'iniziativa'
        pdndAccept:
          type: boolean
          description: >-
            'ENG: Indicates whether the user has correctly accepted the PDND
            Criteria of the initiative - IT: Indica se l'utente ha correttamente
            accettato i PDND Criteria dell'iniziativa'
        selfDeclarationList:
          type: array
          items:
            $ref: '#/components/schemas/SelfConsentDTO'
          description: >-
            ENG: The list of accepted self-declared criteria - IT: Lista dei
            criteri autodichiarati
          maxItems: 10
        userMail:
          type: string
          description: >-
            ENG: Contact email provided by the user during web onboarding - IT:
            Mail di contatto fornita dall'utente in fase di onboarding web
          maxLength: 24
          example: utente@email.it
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
        userMailConfirmation:
          type: string
          description: >-
            ENG: Confirm email provided by the user during web onboarding - IT:
            Mail di conferma fornita dall'utente in fase di onboarding web
          maxLength: 24
          example: utente@email.it
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
...
  parameters:
    ApiVersionHeader:
      name: X-Api-Version
      in: header
      description: 'ENG: Api Version - IT: Versione dell Api'
      required: true
      schema:
        type: string
        enum:
          - v1
        example: v1
        default: v1
  headers:
    Access-Control-Allow-Origin:
      description: >-
        Indicates whether the response can be shared with requesting code from
        the given origin
      schema:
        type: string
        pattern: '^[ -~]{1,2048}$'
        minLength: 1
        maxLength: 2048
    RateLimit-Limit:
      description: The number of allowed requests in the current period
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 240
    RateLimit-Reset:
      description: The number of seconds left in the current period
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 60
    Retry-After:
      description: The number of seconds to wait before allowing a follow-up request
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 240
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer